Added media handler for video urls
The video handler supports youtube, vimeo, dailymotion, vk and html5.
In addition it can setup an iframe with a remote player.

The video code which was located in lightgallery and lg-thumbnail has
been moved to this file in the VideoHandler class.
This should make the video code totally separate from the other
components.